requirements.txt
asgiref==3.6.0
Django==4.1.7
django-js-asset==2.0.0
django-mptt==0.14.0
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0
gunicorn
Pillow
djangorestframework